Fixed editor runtime msf4j issue
